chore: move image edit to its own editor folder
