+++ apps/journeys-admin/src/components/Editor/Drawer/SocialShareAppearance/DescriptionEdit/index.ts
@@ -1,0 +1,1 @@
+export { DescriptionEdit } from './DescriptionEdit'
