+++ apps/journeys-admin/src/components/Editor/Drawer/SocialShareAppearance/TitleEdit/index.ts
@@ -1,0 +1,1 @@
+export { TitleEdit } from './TitleEdit'
