+++ apps/journeys-admin/src/components/Editor/ImageEdit/ImageEdit.spec.tsx
@@ -8,10 +8,10 @@ import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 import {
   GetJourney_journey_blocks_ImageBlock as ImageBlock,
   GetJourney_journey as Journey
-} from '../../../../../../__generated__/GetJourney'
-import { createCloudflareUploadByUrlMock } from '../../../ImageBlockEditor/CustomImage/CustomUrl/data'
-import { listUnsplashCollectionMock } from '../../../ImageBlockEditor/UnsplashGallery/data'
-import { SocialProvider } from '../../../SocialProvider'
+} from '../../../../__generated__/GetJourney'
+import { createCloudflareUploadByUrlMock } from '../ImageBlockEditor/CustomImage/CustomUrl/data'
+import { listUnsplashCollectionMock } from '../ImageBlockEditor/UnsplashGallery/data'
+import { SocialProvider } from '../SocialProvider'
 
 import {
   ImageEdit,
