+++ apps/journeys-admin/src/components/Editor/ImageEdit/Large/Large.tsx
