+++ apps/journeys-admin/src/components/Editor/ImageEdit/Small/index.ts
