+++ apps/journeys-admin/src/components/Editor/ImageEdit/index.ts
@@ -1,0 +1,1 @@
+export { ImageEdit } from './ImageEdit'
